requirements.txt
atomicwrites==1.4.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
coverage==7.5.1
exceptiongroup==1.2.1
gunicorn==19.9.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
more-itertools==10.2.0
packaging==24.0
parse==1.11.1
pluggy==1.5.0
py==1.11.0
pytest==8.2.0
requests==2.31.0
requests-wsgi-adapter==0.4.0
six==1.16.0
tomli==2.0.1
urllib3==2.2.1
WebOb==1.8.5
whitenoise==6.6.0
